add style to upload csv
