+++ portfolio/templates/portfolio/layout.html
@@ -29,6 +29,9 @@
                         <a class="nav-link" href="{% url 'portfolio:load_sell_stock' %}">Sell Stocks</a>
                     </li>
                     <li class="nav-item">
+                        <a class="nav-link" href="{% url 'portfolio:upload_csv_history' %}">Upload History</a>
+                    </li>
+                    <li class="nav-item">
                         <a class="nav-link" href="{% url 'portfolio:view_dividends' %}">View Dividends</a>
                     </li>
                     <li class="nav-item">
