+++ portfolio/views.py
@@ -11,7 +11,7 @@ from asgiref.sync import sync_to_async
 from django.db.models import Q
 from .cron_jobs.process_dividend_payments import process_dividend_payments
 
-from .forms import BuyStockForm
+from .forms import BuyStockForm, CSVUploadForm
 
 import yfinance as yf
 import csv
@@ -62,18 +62,18 @@ class SellStockForm(forms.Form):
             input_formats=["%Y-%m-%d"]
         )
 
-class CSVUploadForm(forms.Form):
-    csv_file = forms.FileField(
-        label='Select a CSV file',
-        help_text='Please upload a CSV file',
-        widget=forms.FileInput(attrs={'accept': '.csv'})
-    )
+# class CSVUploadForm(forms.Form):
+#     csv_file = forms.FileField(
+#         label='Select a CSV file',
+#         help_text='Please upload a CSV file',
+#         widget=forms.FileInput(attrs={'accept': '.csv'})
+#     )
 
-    def clean_csv_file(self):
-        csv_file = self.cleaned_data['csv_file']
-        if not csv_file.name.endswith('.csv'):
-            raise forms.ValidationError('Please upload a valid CSV file.')
-        return csv_file
+#     def clean_csv_file(self):
+#         csv_file = self.cleaned_data['csv_file']
+#         if not csv_file.name.endswith('.csv'):
+#             raise forms.ValidationError('Please upload a valid CSV file.')
+#         return csv_file
 
 # Create your views here.
 @login_required
