+++ static/portfolio/css/form_style.css
@@ -54,6 +54,41 @@
     margin-right: 20px;
 }
 
+.fieldWrapperUploadCsv {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    align-self: baseline;
+    width: 100%;
+    margin-bottom: 10px;
+}
+
+/* Hidden file input */
+#id_csv_file {
+    display: none; /* Hide the default file input */
+}
+
+.label_upload_csv {
+    width: auto; /* Adjusted to auto so it only takes the necessary width */
+    margin: 20px auto; /* Center the button horizontally and add some margin for spacing */
+    padding: 10px 15px;
+    background-color: #054993;
+    color: #f9f7f7;
+    border: 1px solid #054993;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+.file-name {
+    display: none; /* Hide the file name by default */
+    margin-left: 10px;
+}
+
+/* Show the file name when the file input has a value */
+.file-input:not(:placeholder-shown) + .file-name {
+    display: inline;
+}
+
 /* 
 .form-group {
     margin-bottom: 15px;
